+++ packages/typespec-rust/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes
 
 * Optional client method parameters are now in the method's options type and their associated builder methods have been added.
+* Sub-clients now have the suffix `Client` on their type names.
 
 ### Bugs Fixed
 
@@ -13,10 +14,12 @@
 ### Features Added
 
 * Models now derive `typespec_client_core::Model`.
+* Added support for binary responses.
 
 ### Other Changes
 
 * Use macros from `typespec_client_core` for creating enums.
+* `TryFrom` implementations return an `azure_core::Result` instead of `std::result::Result`.
 
 ## 0.1.0 (2024-10-10)