+++ packages/typespec-rust/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Optional client method parameters are now in the method's options type and their associated builder methods have been added.
 * Sub-clients now have the suffix `Client` on their type names.
+* Methods parameters of type `impl Into<String>` have been changed to `String`.
 
 ### Bugs Fixed