+++ packages/typespec-rust/CHANGELOG.md
@@ -13,6 +13,7 @@
 
 * Add necessary calls to `to_string()` for header/path/query params.
 * Fixed improperly clearing an endpoint's query parameters during client construction.
+* Fixed constructing URLs from routes that contain query parameters.
 
 ### Features Added